requirements.txt
Flask==3.0.2
Flask-PyMongo==2.3.0
python-dotenv==1.0.1
gunicorn==21.2.0
dnspython==2.7.0
